Using helper file for duplicated test functions
